Introduction of the MF f functor and M f monad
The idea is that the M f DSL can describe an entire
computation to be carried out by some so called
PFStrategy, which is described by some other monad which
is capable of carrying out PartialFunction get and set.

In this way, the M f term, which describes some data
flow from PartialFunctions, need not pay any mind to the
strategy in which it is evaluated.