Fix race in SmartStrategy
Rather clumsy, but it works.

The problem: it was possible and likely that a resource would be opened
twice but one of the handles left unclosed.

  x <- function someManifest `at_` True
  y <- function someManifest `at_` False

Running this term, the resource for someManifest would be
demanded concurrently by both MAt terms; both would inspect the map and
see that nothing is there, so both would acquire a resource, and whoever
took longer would have the last word and place his resource in the map.
The previously acquired one would go unclosed!